Refactor LoginPage and LoginForm components
